Flask/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.3.3
yfinance==0.2.28
python-dotenv==1.0.0
